+++ ux_valio/facades/named/gstin.py
@@ -34,10 +34,10 @@ class GSTINValidator(StringValidator):
             return False
         return GSTINValidator._luhn_mod_36(value[:14]) == value[14]
 
-    def pre_validation_processing(self, instance: Any, value: Any) -> Any:
+    def _pre_validate(self, instance: Any, value: Any) -> Any:
         if isinstance(value, str):
             value = "".join(value.split()).replace("-", "").upper()
-        return super().pre_validation_processing(instance, value)
+        return super()._pre_validate(instance, value)
 
     def _validate_named_facade(self, instance: Any = None, value: Any = None) -> None:
         if value is None: